refactor: break up release note builder in separate files
- release_notes file was getting too big, this simply break it up
