+++ internal/text/build_heading.go
@@ -1,0 +1,24 @@
+package text
+
+import (
+	"fmt"
+	"strings"
+)
+
+func buildHeading(category string) string {
+	builder := strings.Builder{}
+
+	builder.WriteString("## ")
+
+	heading := fmt.Sprintf("%v ", sectionHeadings[category].title)
+
+	builder.WriteString(heading)
+
+	icon := fmt.Sprintf(":%v:", sectionHeadings[category].icon)
+
+	builder.WriteString(icon)
+
+	builder.WriteString("\n\n")
+
+	return builder.String()
+}
