+++ internal/text/release_notes.go
@@ -1,7 +1,6 @@
 package text
 
 import (
-	"fmt"
 	"strings"
 )
 
@@ -26,50 +25,6 @@ func ReleaseNotes(sections map[string][]Commit) string {
 	if len(sections["others"]) > 0 {
 		builder.WriteString(buildSection("others", sections["others"]))
 	}
-
-	return builder.String()
-}
-
-func buildSection(category string, commits []Commit) string {
-	builder := strings.Builder{}
-
-	builder.WriteString(buildHeading(category))
-	builder.WriteString(buildCommitLog(commits))
-
-	return builder.String()
-}
-
-func buildHeading(category string) string {
-	builder := strings.Builder{}
-
-	builder.WriteString("## ")
-
-	heading := fmt.Sprintf("%v ", sectionHeadings[category].title)
-
-	builder.WriteString(heading)
-
-	icon := fmt.Sprintf(":%v:", sectionHeadings[category].icon)
-
-	builder.WriteString(icon)
-
-	builder.WriteString("\n\n")
-
-	return builder.String()
-}
-
-func buildCommitLog(commits []Commit) string {
-	builder := strings.Builder{}
-
-	for _, commit := range commits {
-		builder.WriteString("- ")
-		// Short version of hash usable on Github
-		builder.WriteString(commit.Hash.String()[:7])
-		builder.WriteString(" ")
-		builder.WriteString(commit.Heading)
-		builder.WriteString("\n")
-	}
-
-	builder.WriteString("\n")
 
 	return builder.String()
 }
